+++ lib/page.rb
