+++ spec/spec_helper.rb
@@ -14,7 +14,7 @@
 #
 # See https://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
 require 'json'
-require_relative '../lib/main'
+require_relative '../lib/page'
 
 RSpec.configure do |config|
   # rspec-expectations config goes here. You can use an alternate
